Fix struct pstore_file_column 64-bit compatibility problem
There's a hole in struct pstore_file_column on 32-bit that's not there on
64-bit which makes files generated on either one of them incompatible with the
other.

Fix that up.

+++ column.c
@@ -16,7 +16,7 @@
 struct pstore_file_column {
 	char			name[COLUMN_NAME_LEN];
 	uint64_t		column_id;
-	uint8_t			type;
+	uint64_t		type;
 	uint64_t		f_offset;
 };
